integrated justeat as connection
